refactor some code, create a relation between the test and the test pod
